# backend-service/doc/swagger.yaml
openapi: 3.0.3
info:
  title: Dell Technologies Resource Management Application Backend API Documentation
  description: >-
    This is the backend API documentation and specifcation for the Dell Technologies Resource Management Web Application, and it is based on the OpenAPI 3.0 specification.


    The API is a RESTful API service which we designed and built to allow us to perform CRUD operations on data stored on a MongoDB Server.


    This application is built using the following technolgoies:

    - [React.js](https://react.dev/)

    - [Node.js](https://nodejs.org/en)

    - [Express.js](https://expressjs.com/)

    - [MongoDB](https://www.mongodb.com/)

    - [Docker](https://www.docker.com/)

    - [Swagger.io](https://swagger.io/)


    This application is built to manage resources much easier 
    and more efficiently helping the organisation to keep 
    track and manage resources more easily. 


    It is built using the MERN stack and it is a full-stack 
    application which can be easily deployed with the help 
    of Docker deployment. We also utilised React.js as a 
    frontend framework which works well with Express.js and 
    Node.js as the backend server. For the database, we have
    used MongoDB, since it is a NoSQL database, it allows 
    for much more flexibility as well as scalability.
    Lastly, for the ease of testing and documentation, we 
    have utilised Swagger.io to document our API endpoints.


...
    below.


...

...
  contact:
...
  license:
...
servers:
...
paths:
...
  /admin/getworkshop:
    get:
      tags:
        - admin
      summary: Get all workshop requests
      description: Get all workshop requests
      operationId: getWorkshopRequests
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkshopRequest"
        "401":
          description: No authorization token provided
        "403":
          description: Access forbidden
        "500":
          description: "Error Retrieving Workshops: Internal Server Error"
      security:
        - admin_auth_token: []
  /admin/getworkshop/{workshopId}:
    get:
      tags:
        - admin
      summary: Get a workshop request by ID
      description: Get a workshop request by ID
      operationId: getWorkshopRequestById
      parameters:
        - name: workshopId
          in: path
          description: ID of WorkshopRequest to return
          required: true
          schema:
            type: string
            description: MongoDB ObjectID as a hexadecimal string.
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkshopRequest"
        "401":
          description: No authorization token provided
        "403":
          description: Access forbidden
        "404":
          description: Workshop not found
        "500":
          description: "Error Retrieving Workshop: Internal Server Error"
      security:
        - admin_auth_token: []
  /admin/addworkshop:
    post:
      tags:
        - admin
      summary: Add a new workshop request
      description: Add a new workshop request
      operationId: createWorkshopRequest
      requestBody:
        description: Create a new workshop request
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkshopRequest"
          application/xml:
            schema:
              $ref: "#/components/schemas/WorkshopRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/WorkshopRequest"
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkshopRequest"
            application/xml:
              schema:
                $ref: "#/components/schemas/WorkshopRequest"
        "401":
          description: No authorization token provided
        "403":
          description: Access forbidden
        "500":
          description: "Error Creating Workshop: Internal Server Error"
      security:
        - admin_auth_token: []
  /admin/updateworkshop/{workshopId}:
    put:
      tags:
        - admin
      summary: Update a workshop request by ID
      description: Update a workshop request by ID
      operationId: updateWorkshopRequest
      parameters:
        - name: workshopId
          in: path
          description: ID of WorkshopRequest to update
          required: true
          schema:
            type: string
            description: MongoDB ObjectID as a hexadecimal string.
      requestBody:
        description: Update an existent workshop request
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkshopRequest"
          application/xml:
            schema:
              $ref: "#/components/schemas/WorkshopRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/WorkshopRequest"
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkshopRequest"
            application/xml:
              schema:
                $ref: "#/components/schemas/WorkshopRequest"
        "401":
          description: No authorization token provided
        "403":
          description: Access forbidden
        "404":
          description: Workshop not found
        "500":
          description: "Error Updating Workshop: Internal Server Error"
      security:
        - admin_auth_token: []
  /admin/deleteworkshop/{workshopId}:
    delete:
      tags:
        - admin
      summary: Delete a workshop request by ID
      description: Delete a workshop request by ID
      operationId: deleteWorkshopRequest
      parameters:
        - name: workshopId
          in: path
          description: ID of WorkshopRequest to delete
          required: true
          schema:
            type: string
            description: MongoDB ObjectID as a hexadecimal string.
      responses:
        "200":
          description: successful operation
        "401":
          description: No authorization token provided
        "403":
          description: Access forbidden
        "404":
          description: Workshop not found
        "500":
          description: "Error Deleting Workshop: Internal Server Error"
      security:
        - admin_auth_token: []
...
  /workshop/:
    get:
      tags:
        - workshopRequest
      summary: Get all workshop requests
      description: Get all workshop requests
      operationId: getWorkshopRequests
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkshopRequest"
        "500":
          description: "Error Retrieving Workshops: Internal Server Error"
      # security: [] (authentication required for admin)
  /workshop/get/{workshopId}:
    get:
      tags:
        - workshopRequest
      summary: Get a workshop request by ID
      description: Get a workshop request by ID
      operationId: getWorkshopRequestById
      parameters:
        - name: workshopId
          in: path
          description: ID of WorkshopRequest to return
          required: true
          schema:
            type: string
            description: MongoDB ObjectID as a hexadecimal string.
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkshopRequest"
        "400":
          description: Invalid ID supplied
        "404":
          description: Workshop not found
        "500":
          description: "Error Retrieving Workshop: Internal Server Error"
      # security: [] (authentication required for admin, trainer and client)
  /workshop/add/:
    post:
      tags:
        - workshopRequest
      summary: Add a new workshop request
      description: Add a new workshop request
      operationId: createWorkshopRequest
      requestBody:
        description: Create a new workshop request
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkshopRequest"
          application/xml:
            schema:
              $ref: "#/components/schemas/WorkshopRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/WorkshopRequest"
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkshopRequest"
            application/xml:
              schema:
                $ref: "#/components/schemas/WorkshopRequest"
        "400":
          description: Invalid input
        "422":
          description: Validation exception
        "500":
          description: "Error Creating Workshop: Internal Server Error"
      # security: [] (authentication required for admin and client)
  /workshop/update/{workshopId}:
    put:
      tags:
        - workshopRequest
      summary: Update a workshop request by ID
      description: Update a workshop request by ID
      operationId: updateWorkshopRequest
      parameters:
        - name: workshopId
          in: path
          description: ID of WorkshopRequest to update
          required: true
          schema:
            type: string
            description: MongoDB ObjectID as a hexadecimal string.
      requestBody:
        description: Update an existent workshop request
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkshopRequest"
          application/xml:
            schema:
              $ref: "#/components/schemas/WorkshopRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/WorkshopRequest"
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkshopRequest"
            application/xml:
              schema:
                $ref: "#/components/schemas/WorkshopRequest"
        "400":
          description: Invalid input
        "404":
          description: Workshop not found
        "422":
          description: Validation exception
        "500":
          description: "Error Updating Workshop: Internal Server Error"
      # security: [] (authentication required for admin)
  /workshop/delete/{workshopId}:
    delete:
      tags:
        - workshopRequest
      summary: Delete a workshop request by ID
      description: Delete a workshop request by ID
      operationId: deleteWorkshopRequest
      parameters:
        - name: workshopId
          in: path
          description: ID of WorkshopRequest to delete
          required: true
          schema:
            type: string
            description: MongoDB ObjectID as a hexadecimal string.
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid input
        "404":
          description: Workshop not found
        "500":
          description: "Error Deleting Workshop: Internal Server Error"
      # security: [] (authentication required for admin)
...
components:
  schemas:
...
    WorkshopRequest:
      type: object
      properties:
        _id:
          type: string
          example: 668d5ba4a5f8eca9ff47283e
          description: MongoDB ObjectID as a hexadecimal string.
        name:
          type: string
          example: Introduction to Dell Technologies
        description:
          type: string
          example: "Workshop about Introduction to Artificial Intelligence and Machine Learning."
        startDate:
          type: string
          example: 12th July 2024
          description: Moment.js object for the start date of the workshop.
        endDate:
          type: string
          example: 14th July 2024
          description: Moment.js object for the end date of the workshop.
        location:
          type: string
          example: Local/Overseas
        timeStart:
          type: string
          example: 9:00 AM
          description: Moment.js object for the start time of the workshop.
        timeEnd:
          type: string
          example: 5:00 PM
          description: Moment.js object for the end time of the workshop.
        duration:
          type: integer
          example: 10
          description: Difference of days between the start and end date of the workshop.
        status:
          type: string
          example: pending
          default: pending
          enum:
            - pending
            - approved
            - rejected
        type:
          type: string
          example: workshop
          enum:
            - workshop
            - recreational
            - talks
            - educational
        maxParticipants:
          type: integer
          example: 20
        client:
          type: string
          example: 668d5ba4a5f8eca9ff47283e
          description: MongoDB ObjectID as a hexadecimal string.
        trainer:
          type: object
          properties:
            _id:
              type: string
              example: 668d5ba4a5f8eca9ff47283e
              description: MongoDB ObjectID as a hexadecimal string.
      xml:
        name: workshop
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    admin_auth_token:
      type: http
      scheme: bearer
      bearerFormat: JWT
    trainer_auth_token:
      type: http
      scheme: bearer
      bearerFormat: JWT
    client_auth_token:
      type: http
      scheme: bearer
      bearerFormat: JWT
    api_key:
      type: apiKey
      name: api_key
      in: header